Telephony v2.6 porting
To fix missed part of v2.6 porting.
Mainly refactor for commit 2449ca524270e1dc744d52a9db0bb594d2be5c29

+++ dom/telephony/TelephonyCall.h
@@ -139,7 +139,7 @@ public:
   IMPL_EVENT_HANDLER(groupchange)
 
   static TelephonyCallState
-  ConvertToTelephonyCallState(uint32_t aCallState);
+  ConvertToTelephonyCallState(uint16_t aCallState);
 
   static TelephonyCallVoiceQuality
   ConvertToTelephonyCallVoiceQuality(uint16_t aQuality);
